requirements.txt
streamlit
numpy
opencv-python
pytesseract
pillow